Move remaining unit tests into JsonApiDotNetCoreTests project
