+++ src/JsonApiDotNetCore.Annotations/Properties/AssemblyInfo.cs
@@ -4,5 +4,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Benchmarks")]
 [assembly: InternalsVisibleTo("JsonApiDotNetCore")]
 [assembly: InternalsVisibleTo("JsonApiDotNetCoreTests")]
-[assembly: InternalsVisibleTo("UnitTests")]
 [assembly: InternalsVisibleTo("TestBuildingBlocks")]
