+++ src/JsonApiDotNetCore/Properties/AssemblyInfo.cs
@@ -3,4 +3,3 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Benchmarks")]
 [assembly: InternalsVisibleTo("DapperExample")]
 [assembly: InternalsVisibleTo("JsonApiDotNetCoreTests")]
-[assembly: InternalsVisibleTo("UnitTests")]
