+++ test/JsonApiDotNetCoreTests/UnitTests/AssemblyTypeLocator/DerivedType.cs
@@ -1,0 +1,3 @@
+namespace JsonApiDotNetCoreTests.UnitTests.AssemblyTypeLocator;
+
+internal sealed class DerivedType : GenericBaseType<int>;
