+++ test/JsonApiDotNetCoreTests/UnitTests/AssemblyTypeLocator/GenericBaseType.cs
@@ -1,0 +1,4 @@
+namespace JsonApiDotNetCoreTests.UnitTests.AssemblyTypeLocator;
+
+// ReSharper disable once UnusedTypeParameter
+internal class GenericBaseType<T>;
