+++ test/JsonApiDotNetCoreTests/UnitTests/AssemblyTypeLocator/IGenericInterface.cs
@@ -1,0 +1,4 @@
+namespace JsonApiDotNetCoreTests.UnitTests.AssemblyTypeLocator;
+
+// ReSharper disable once UnusedTypeParameter
+internal interface IGenericInterface<T>;
