+++ test/JsonApiDotNetCoreTests/UnitTests/AssemblyTypeLocator/InterfaceImplementation.cs
@@ -1,0 +1,6 @@
+using JetBrains.Annotations;
+
+namespace JsonApiDotNetCoreTests.UnitTests.AssemblyTypeLocator;
+
+[UsedImplicitly(ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature)]
+internal sealed class InterfaceImplementation : IGenericInterface<int>;
