+++ test/JsonApiDotNetCoreTests/UnitTests/AssemblyTypeLocator/TestResource.cs
@@ -1,0 +1,5 @@
+using JsonApiDotNetCore.Resources;
+
+namespace JsonApiDotNetCoreTests.UnitTests.AssemblyTypeLocator;
+
+internal sealed class TestResource : Identifiable<long>;
