+++ test/JsonApiDotNetCoreTests/UnitTests/Resources/IdentifiableTests.cs
@@ -2,7 +2,7 @@ using FluentAssertions;
 using JsonApiDotNetCore.Resources;
 using Xunit;
 
-namespace UnitTests.Models;
+namespace JsonApiDotNetCoreTests.UnitTests.Resources;
 
 public sealed class IdentifiableTests
 {
