--- test/UnitTests/Graph/BaseType.cs
@@ -1,5 +1,0 @@
-// ReSharper disable UnusedTypeParameter
-
-namespace UnitTests.Graph;
-
-internal class BaseType<T>;
