--- test/UnitTests/Graph/DerivedType.cs
@@ -1,3 +1,0 @@
-namespace UnitTests.Graph;
-
-internal sealed class DerivedType : BaseType<int>;
