--- test/UnitTests/Graph/IGenericInterface.cs
@@ -1,5 +1,0 @@
-// ReSharper disable UnusedTypeParameter
-
-namespace UnitTests.Graph;
-
-internal interface IGenericInterface<T>;
