--- test/UnitTests/Graph/Implementation.cs
@@ -1,6 +1,0 @@
-using JetBrains.Annotations;
-
-namespace UnitTests.Graph;
-
-[UsedImplicitly(ImplicitUseKindFlags.InstantiatedNoFixedConstructorSignature)]
-internal sealed class Implementation : IGenericInterface<int>;
