--- test/UnitTests/Graph/Model.cs
@@ -1,5 +1,0 @@
-using JsonApiDotNetCore.Resources;
-
-namespace UnitTests.Graph;
-
-internal sealed class Model : Identifiable<long>;
